Change int id to long

+++ src/main/java/com/example/myapp/models/User.java
@@ -2,7 +2,7 @@
 package com.example.myapp.models;
 
 public class User {
-    private Integer id;
+    private Long id;
     private String username;
     private String password;
     private String firstname;
@@ -13,7 +13,7 @@ public class User {
     private String email;
 
     // Create a new User
-    public User(Integer id, String username, String password, String firstname, String lastname, String dob, String role, String email) {
+    public User(Long id, String username, String password, String firstname, String lastname, String dob, String role, String email) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -25,7 +25,7 @@ public class User {
     }
 
     // Create a new User with basic info
-    public User(Integer id, String username, String password, String role) {
+    public User(Long id, String username, String password, String role) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -36,11 +36,11 @@ public class User {
         this.email = "";
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
